airflow/requirements.txt
psycopg2-binary
polars
structlog
apache-airflow-providers-docker
docker
fuzzywuzzy